fixed LanguageSelect props, reorganized component imports because I changed file structure of components folder
