+++ src/app/[locale]/page.tsx
@@ -1,6 +1,6 @@
 import { HeroSection } from '@/components/landing/HeroSection'
 import { HowItWorksSection } from '@/components/landing/HowItWorksSection'
-import { Container } from '@/components/Container';
+import { Container } from '@/components/layout/Container';
 import { TraditionSection } from '@/components/landing/TraditionSection'
 
 export default function Home() {
